Hysen HY08ACF: minor additions
Based on config from last comment on #343

- add product id
- add Holiday mode

The comment alludes to a problem with the detection, but no dp
types seem to be different in the posted config and what is currently
installed.  hvac_action was renamed to valve, which removes
functionality, and some values that other users are reporting were
removed from the mapping, but since this is a read-only attribute,
there should not be any harm from leaving them in. There was also a duplicate
of fan speed into its own select entity, but this is also not needed, as it
is already accessible through the climate entity.

Issue #343