Add support for Konlen/Rockson WF96L water level controllers.
Issue #381

There seem to be two variants of this being widely sold.
Rockson with Chinese language labels on the buttons.
Konlen with English language labels on the buttons.

I chose to label this with the Konlen branding, but it probably
applies to both, and the original report doesn't specify which.