Smartplugs (multiple): do not mark as energy or total_increasing
Also applies to a couple of heatpumps and thermostats that had power
consumption readings.

Due to differences in the way different Tuya devices handle energy
dps (add_ele in particular), it is not possible to get consistent
behavior without warnings in the log.  Even in devices that seem to
work, the energy reported by HA is off by 5 - 10% compared to the Tuya
reports, so it is probably better to use an Integration helper on the
power sensor rather than using the energy sensor directly.

Left dedicated power meters alone, as those seem to report a running total
as expected.

Issue #392