Meters: Standardize energy sensors.
Energy sensors must be either total_increasing, or total with
last_reset.  Since we don't support history tracking to detect reset
and save the time, only total_increasing will work.

For Balance (on prepay meters), it is actually always decreasing, but
is not really a measurement of energy, it is credit remaining, so
don'r mark those as energy.