Config flow: migrate unique_id of entities to include platform.
When a device contains two entities of different types with the same
name,  an ERROR is logged about duplicate unique IDs, but things seem
to still work, as the config_id is different.  Probably customizing
these entities through the UI will not work however.

If we include the entity platorm in the unique_id, we can eliminate
these duplicates.

Issue #455