Add support for a 4-way power monitoring powerstrip.
Issue #457

Although the reported issue is for a 3-way + USB device, the dp layout
follows the standard layout for a 4-way (USB is dp 4, not 7 as is
often assigned to USB).  Timer and power monitoring layout is also
consistent with the standard most multi-switch devices seem to follow,
so this makes a better generic 4-way than the existing
"energy_monitoring_powerstrip" config, which uses non-standard dp ids
for the power monitoring, and does not support timers.