Camera support: BCom intercom camera.
This commit adds a still image capable camera.  There is not yet
support for streaming, or cameras that provide still images via linked
URLs only.  It is not tested against the actual camera yet, so may
turn out not to work.

Issue #371, potential for use for #168, though the user there
documented the still image as being a URL (but that may be for the cloud API).