smartplugv2_energy: make energy and test_bit dps optional
Energy is sent infrequently by many plugs, so don't require it.
Test bit is not really needed, like calibration dps, so don't require it.

The only definite difference now between this and the simpler smartplugv2 is the
Initial State selector.  Overcharge cutoff and Energy sensor are still there,
but optional.