smartplugv1: don't force updates of power/current/voltage
These older plugs do not use the standard dps of 18,19,20 for these, and
there don't seem to be any reports of them needing the workaround.

Localtuya seems to only enable updatedps by default for plugs using 18,19,20
as the dps ids for these, and running protocol v3.3 or 3.4.

updatedps seems to cause problems for Gosund EP2 plugs which use this dp
layout with protocol v3.3

Issue #477