Add support for device class to number platform.
This was silently added to HA in 2023.2.0.  Unfortunately it is
missing duration, which is the most common use for numbers, and both
sensor and number are missing a class for relative temperatures, which
is the second (using temperature would cause incorrect conversions
due to the misalignment of 0 points between Celsius and Fahrenheit).

So only a few instances of temperature limits, and some power,
current, voltage and illumanance configs can be given device classes.

Issue #513