Snap numeric values to the nearest mapping.
Some configs have been made with numeric values with mappings, to mtch
specific devices.  If step was set and the mapped values evenly
spaced, this works for things like temperature where step is used by
HA.  But in the general case, it does not work.  Some of the existing
configs for light brightness with fixed steps assumed it works in the
general case, so were broken.

- Aspen ADV200: map brightnesses to 20%, 50&, 100% per the manual,
  instead of evenly space 33%, 66%, 100%.
- Aspen tests: as this light does not have an off setting, modify the
  tests to allow that, instead of relying on garbage in-garbage out,
  which no longer works with it snapping to nearest valid value.

- Heatstorm HS6000GC: invert the child lock per the device behaviour

Issue #510