Arlec thermostat: change switch from dp 1 to dp 2.
Based on diagnostics, it looks like there is never a dp 1 exposed by the device.
Probably though dp 2 works as a manual switch regardless of whether it is in
auto or manual mode, so make that a switch, even though it partially duplicates
the hvac_mode of the climate entity.

Issue #535