TOMPD 63LW: change "energy reset" to a boolean.
string (with pushed value "empty") was based on iot.tuya.com info, with no
log info about this dp to corroborate it.

Now there is a report about an almost identical device which does have log
info for this dp, the only mismatch with this config being that the energy
reset is a boolean.  After confirming back with the TOMPD 63LW user, energy
reset did not seem to be working, so likely it is actually boolean there also.

Issue #577, #388