Moebot: mark optional dps, change password to integer
Original config was based on info from another project, but string/integer
distinction wasn't clear, and it wasn't obvious which dps were always
returned, and which were optional.  Now we have a log message we can refine
things so the device can be matched (seems to require explicit protocol
version setting of 3.4).

Issue #573