Revert removal of "auto" protocol version.
Subdevice commit removed the possibity to set "auto" as protocol
version.  This is not friendly to end users, who should not be
expected to know the internal details of their device, or to do the
trial and error to find the correct protocol themselves.

Likely a change is needed to set the correct protocol on parent
devices when the protocol is changed, and that is the change that
should be made to fix the issue, not removal of the feature like this.

Partial revert of PR #651