cover: remove special handling of reversed dp.
When this was added, I thought that this dp would indicate when the position
should be flipped.  But now that a few more covers have been implemented, it
is clear that Tuya's idea of position is always the reverse of HA's, and this
is just a switch for the motor direction which doesn't affect position
reporting.

Update configs that used this to always invert positions.