Cover: refactor state logic into a single function.
is_opening, is_closing and is_closed had independent logic, so there
was a possibility that they were returning inconsistent values.  Use a
single function to work out the current state of the cover based on
whatever information is available, returning None if there is any
ambiguity, to avoid leaving it stuck in a state.  Modify the API
functions to use the new function.

- update tests for some of the simpler devices, as the state now
returns None in more cases.

- mark zemismart_roller_blinds action dp as unreliable. Leave it for
monitoring, but don't use it to determine cover state.

Trying to resolve the issues raised in PR #672, #655 without adding
internal state, which is likely to get out of sync with the actual state.