Energy meters: standardise units to A, kW
For energy meters, the device's intention is to monitor a whole
circuit, not just a single device which may be low power.  So it makes
sense for such devices to return current in amps, and power in kW,
even if the underlying device is giving mA and 0.1W precision like most
smartplugs do.

This is a frequently asked change, which can be taken care of in the
UI, but making the change at least for energy meters should head off
further requests.

For smartplugs, a nore convincing case can be made that it is a matter
of preference and that many use cases will be for loads where the
smaller units make more sense.