Use HS color rather than RGBW for colour lights.
The initial implementation of lights made the mistake of choosing
RGBW, thinking that the W would correspond to the brightness dp and
RGB to colour.

After implementing though, brightness dp should not be set for colour
mode, and the lights have either RGBHSV settings or HSV in the colour
dp.  Since HSV is the common factor, HS color fits better (V is taken
from brightness).

This changes the UI to remove the confusing and non-working "Color
brightness" and "White brightness" sliders from the color
tab. Brightness on the main tab still works.

Issue #691