Use brightness to turn on light when there is no other way.
We used to have this feature, but it was lost in some of the refactoring,
I think when a light that did not support turning off was added.

Note though that the light will always come on at full brightness if we
rely on the brightness control to turn on and off.

Should solve the issue mentioned in PR #743