ZX-M100 presence sensor: sensor is string
The main sensor is a string "presence" or "none", not a boolean.
Looking back at the original report, the information was there then also,
and no log was received showing a boolean, nor was there confirmation from
the reporter that it was working.  I think the assumption was made based on
another device that had recently been encountered that a two option enum in
the cloud protocol would become a boolean in the local protocol.

Also, the distance is missing, as in the original report it was not identified.

Issue #755 (original was #636)