Goldair GPPH heater: map power level of auto to 0
GPPH 620 returns "auto" as a power level, others in the series seem to
return the actual output level when in auto mode.  Since we cannot guess
what the power level might be, map it to 0.  It may even only do this when
an actual power level is not available due to the heater being off.

Issue #762