feat(device): better debug around why device doesn't match config
When an end-user attempts to add a device, this change adds some
more detailed logging around why the device doesn't match a known
config. Details of the required DPs that are missing, or the DPs with
unmatched type are printed.

This routine is hit when an end-user attempts to add a device, but not
during normal operation. The decreased performance in the matching
routine should be acceptable given the improved information available
and the low frequency of use.

The end-user will need to enable debug-level logging for the component to view this level of detail, so it won't clutter the logs.