HCT611: Add additional smart weather state, use HA weather strings
Device has "cloud" as well as "cloudy" according to comment on #747.
Map the device weather strings to the underlying strings of HA weather platform.
This will allow using the translations from there when we can figure out how
to do that, and also should allow passing values directly from a weather
integration to the select to show on the device (if cloud does not conflict
with local control of that).