Linkoze switch: tweak values
Contrary to the portal docs, the log is indicating "single_click", not "click"
There is also a "double_click" state that I missed.

It seems to report the last action persistently rather than clear the state,
so maybe the persist: false was not needed, but in case it is, use it
consistently.

Issue #774