Rearrange SonarCloud scan conditions.
Replace hack that checked out main (regardless of what branch the request was for) before the SonarCloud scan with a condition to only run SonarCloud on main branch.
Add check that the repository is make-all/tuya-local, as me pushing rebases to
pull requests is triggering the scan, which then runs without a token since it is using the source repository config.