Covers: duplicate position as current_position for some devices.
Some cover devices let position be set via the position dp, but leave
it showing the last manually set position even when the cover has
moved subsequently.  For this reason, the cover was modified not to
trust position when determining if the cover is open or closed, since
reporting either of those incorrectly results in the corresponding
control being disabled in the UI when it shouldn't be.

However some devices do report the current position reliably using the
same position dp used to set position, and with the new code those
devices have regressed in functionality.  In order to get that
functionality back, we need to mirror the same dp as current_position
as well as position.

Issue #715, #794

As devices using simple_blind.yaml are more generic, they may be of
either of the above types, so a second config was made, allowing the
user to try both.