Maintain order when deduplicating values.
When returning the list of possible values, we removed duplicates by
the fastest method in Python of converting to a set then back to a
list.  However, the conversion to a set loses the order of items.
Instead construct a new list by iterating through them discarding ones
that have been seen before.

Issue #808