Mark fans with translation_key fan_with_presets
Based on the hassfest errors, it seems we cannot append to default
translations, so we need to use a translation_key to add translations
to the fan preset_mode values.

As the upstream does not have any translations, I suspect they have
not yet got to fans, and in future this may be able to be partially
reverted, or there may be more disruption when they choose different
presets (nature and sleep seem common enough that they will be
included, maybe with some variation on nature).

Issue #682