Docs:  iot.tuya.com API Explorer has been reorganised.
Change references to API functions to reflect the new locations.
The old function to get a single device's information that worked for
getting product id and local key seems no longer available, but the
bulk function is available.  Both that and the function for listing dp
ids have gotten new names in the API Explorer and been shuffled around
the heirarchy.  Overall the new heirarchy is an improvement, with less
duplication and no confusingly similar names.

Discussion #823, thanks to @longavailable for the heads up.