Merge Lytmi Fantasy3 and Lytmi Fantasy/Neo 3 configs.
The newer config is just an improvement of the older one, with more
unknown dps identified, and the missing color_mode dp removed.  Since
these are actually for the same device, they would be better as a
single config, with the older name taking preference for backwards
compatibility snice the newer one is not yet released.

Add docs for this and other recent PRs.