fan: turn off when speed is set to 0 and on when speed set otherwise
Speed 0 turning the fan off used to be the behaviour of HA, but it
changed recently, probably becaue fans exist somewhere that really
have a 0 speed.  Many of our configs give a minimum speed of 0 when 0
is not a valid speed step, so we are relying on this implicit
behavior.  I think HA anyway doesn't check the range of speeds, only
the number of steps, and always offers Off as an option.

Also, the new UI buttons result in the on switch being hidden when
more info is open, even though inclusion of Off in the speed buttons
implies that it is possible to switch the fan off and on from here.
So the reverse behaviour of implicitly turning the fan on when speed
is set is also implemented by this commit.

Issue #936