Remove names from sensors with device class matching name.
Issue #871

Now that device class is used as an alternative to name to construct
unique ids, the names can be removed from more entities to allow the
default class based names to be used along with any translations.

This commit removes only names that are an exact match for the class.
Other candidates like PM2.5 are left for a later iteration once
migration has been figured out, as the stub version pm2_5 differs from
the class pm25, and we need to deal with the migration issues due to
class being used in the unique id where the name was previously
missing (mostly primary entities) before making the problem worse.