Catit pixi smart feeder: Add optional binary sensors for fault states
Original PR #757 did not mention dp 14, but PR #960 for seemingly compatible
Nutri Meal has it in the log, and in the device description along with some
dps 101 and up which do not appear in the log, and are not added.
These are optional, so their absense will just report no problems if the
original Catit device does not support the fault reporting.