Migrate entity ids to use device_class when unnamed.
Some older entity ids will be in the entity registry with just the
platform, but if they are marked with a device_class, the id will now
use that.  The registered entities need to be migrated to match the
new ids to avoid leaving zombie entities in the registry, and losing
any user customisations of name, icon etc.  It may also help with
finding the right entity in automations etc, but I am not sure so will
still list the possiblity that automations will need updating in
breaking changes.

Issue #871