Eberg Cooly C35HD: make temp_set_f optional
It seems this dp is not always included when temp is set to C.

Be Cool has the same dp layout, but different mode mapping, and is
detected for Eberg devices, the only differences being a lack of
select entity to set the units, and the temp_set_f is already marked
as optional.  Sync these two by also adding the select entity for
temperature selection to Be Cool.

Remove the test for BeCool rather than update for this change, as it
is not testing anything unique.

Issue #974