Revert to overriding entity name property
Although the logic to use translated entity names is in the superclass
name implementation, setting _attr_name to None results in
use_device_name returning true, as that is the first condition
checked, and the device name is used instead of the default entity
name.  This is probably an HA bug, as it becomes quite complex to use
default name translations as a result, and certainly isn't per the
documentation.

Issue #871