Do not reverse map hidden values.
A common pattern for optional dps is that we want to map null to a default
value so when it is not reported, we can report a value.
To avoid such a mapping showing up as an option that can be set, we can mark
these mappings as hidden.  But also, these mappings should not be reversed
when setting the value:
Consider:
 - a switch has values on and off (True/False), which we just pass through.
   But the dp is optional, and we want to show it as off when missing, so
   we add a mapping from null to false.  With the previous code, every time
   we turn the switch off from HA, the mapping would be applyed in reverse,
   and we would try to set the dp to null instead of false.  By marking the
   mapping as hidden, we can now avoid that reverse mapping.