Kogan LX10 vacuum: rearrange commands to work more reliably
- rename start dp to activate, which has special handling in code.
- add a conditional mapping based on pause dp.

Vacuum: reorder handling of activate and commands so that commands
take priority.  This allows pause to be handled by command, while
activate dp is used for start, so that the limitation of a single
conditional constraint can be worked with when start and pause are
separate binary dps from the command.

PR #981