Avatto curtain+light: rearrange optional dps
Latest log shows light present but calibration missing.
This is more reasonable, as it would allow to distinguish from other
curtain-only devices still.  Since the device is obviously not
reliable in which dps it decides to leave out, users are probably
going to have to retry a few times to get it recognised.

Issue #987