Label Sound type according to Tuya app.
The iot.tuya.com docs seemed to suggest that this was used together
with dp 10, so the assumption was it selects between different groups
of 34 sounds.  But the app screenshot shows only 11 sounds selecatable
within one category, so it may just be filtering the list of 34, and
therefore redundant.

Since I don't know for sure which of these, it is left in for now, and
changed to match the app naming.

Issue #1042