Add support for another variant of Liectroux XR500 vacuum.
Previously Liectroux XR500 was found to match Kabum Smart 700, but
another user has a variant that almost matches Abir X8, which is
identified as having a "T3" motherboard.

This differs from the Abir X8 in having additional "Off" options
for suction and mopping (presumably to use the features separately),
additional "remaining life" sensors and reset buttons for all the
consumables, and is missing the "cliff detect" config switch.

Issue #1056