HA 2023.9.0 has altered the config flow return values.
Config flow tests need updating to include the new "preview" entry
that is being returned.

Update version required to ensure older versions that do not have this
are not used for the tests.  Since we don't make use of it, this does
not affect the version of HA needed by end users at runtime, only the
automated tests.

- minor update to Ditua config: HSV ranges should start from 0 not 1.