ElectriQ Airflex: provide single DP versions of hvac_mode
The original config for this model also had an Eco preset, so the
original config used preset_mode mostly duplicating hvac_mode.  After
it was discovered that Eco mode did not really exist, the preset_mode
was changed to a hidden work_mode dp, leaving just the hvac_mode
controls.

But apparently at least some devices do not work like this.
Originally the reports were that hvac_mode only worked for on/off, it
didn't change the mode, but after the change d682aff2519, the reverse
became true.

A proper fix is probably to allow labelling the config as "single dp
only" and implementing queuing in the sending code to ensure only one
dp is sent at a time, but the quick fix is to restore the preset_mode
duplication, and add a separate on/off switch, which can be used as
alternatives to hvac_mode.

Discussion #609