Change light color temperature from mireds to Kelvin.
Use the new color_temp_kelvin API for lights instead of the
old mired API which was apparently deprecated in November.

Implement min and max by setting the _attr variables in the base class
when initialising.  To support this, added a new general
`target_range` mapping to replace the hardcoded shifting of the
color_temp range into the 152 - 500 mired range (2000-6500K).

Configs using color_temp were updated by removing the `invert: true`
mapping, and adding a `target_range`.  Where no information about the
range could be found from a quick web search, 2700-6500 was used, as
this seems to be most common for full-range CCT bulbs.

The following devices did not have the color_temp inverted.  It is not
certain whether this is because they are following a mired scale,
or this was a bug that hadn't been noticed.  Since all known Tuya
bulbs follow the Kelvin scale, no change was made to these to invert
them (assuming it was a bug that they were not inverted before).

 - atomi_ceiling_fan
 - rgbcw_lightbulb2
 - tampa_led_system
 - windcalm_fan_with_light

The following two had some evidence that the lack of inversion was
deliberate, so the `invert: true` was added to maintain that.

 - arlec_fan_light
 - fanco_ecosilentdeluxe

Issue #1086