Smartplugv2 Energy: add compatible Eightree device
Based on that device, add optional child lock and light control, which are
reasonably common on this type of device and do not conflict with other
known matches.

Remove polling, as it is known to cause problems, and this is close enough
to smartplugv2 that smartplugv2_polled can be used if it is needed.

Issue #1061