Treat empty cid the same as no cid
When processing user entry through the config flow, we get
an empty string rather than None for an empty cid field.
This causes the test connection to fail when reconfiguring
(but not the original connection for some reason, otherwise
this would have been noticed sooner than 6 months after the
introduction of cid support).

Issue #1110