Add support for Teberno LED lights
These are a bit problematic, as the logs only show them with dps
20 (switch) and 26 (timer), but they have rgb and scene dps listed in
the tuya docs, as well.  Home Assistant seems to have an assumption
that any light capable of more than just on/off also supports
brightness, which does not seem true of this light in white mode.
It is possible that white mode is not really supported, but then the
initial colour setting is unknown...

Since the log did not show the RGBHSV format, and the
QueryThingsDataModel result did not break down the fields, I had to
make a guess that with this dp layout it will follow the newer 6 byte
hsv format, rather than the older 7 byte rgbhsv.

Issue #1095