Use HA entity registry to disable deprecated entities by default.
Previously our device config had a list of checkboxes for each entity
to allow enabling and disabling in the integration config.
But these were removed when HA made it easy to enable and disable
entities from the main Settings / Devices UI.  Some code related to
these was still present, causing deprecated entities to be disabled by
default, but needing the old config UI to enable them.

Instead, use the entity_registry_enabled_default property to indicate
that it should be disabled in the registry by default, but otherwise
set it up as usual.

Issue #1143