Inkbird ITC-306A: use hvac_mode for control instead of preset_mode.
hvac_mode allows the device to be toggled on and off, and is easier to
use in the UI.  on and off are simply mapped to "off" and "heat", but
pause is creatively mapped to "fan_only" to allow that to be selected
(according to user feedback it is no different than off mode, but
according to the manual, pause should leave the display on for
monitoring temperature, while off should be completely off).

- modify the translations for "thermo_switch" (only used by this
device) to follow this change.

- modify hvac_action to indicate "off" for both off and pause modes.

Discussion #1144