ZY-M100: make status sensor optional.
WZ-M100 has the same dps mapping, except for the status sensor.
Make that optional, so the config can work for both devices.
Adjust step of fading and detection time (assume the ZY device is
the same, as going through the complete range with a step of 5 would
take forever).

Issue #1132