Lock: clear non-persistent unlock credential dps on reporting.
When the last credential is kept until the next full poll, if the user
unlocks twice in quick succession, the second attempt is not reported.
This can also happen with a different credential if a lower priority
unlock method is used

Clearing immediately after reporting may mean it disappears from the
attributes too quickly to notice in the UI, but the logbook will be more
accurate.

Issue #796