climate: revert back to returning auto for no hvac_mode
When we return None, the entity card shows a status of "Unknown".

When we return "auto", in addition to the mode "Auto", the entity card
also shows set temperature, current temperature and hvac_action
if available.

The latter is definitely better, even if the mode is wrong.

To resolve the issue of an empty mode select box, include "auto"
in the hvac_modes, as the card may be checking for only one entry
in the list rather than one or none to hide this select box.

Issue #1144.

Reverts 82fa229ee7cb79