match_against util: better indicate match failures
There was no distinction between dps missing that were optional,
and those that were required.  Output an extra message when
required dps are missing in addition to noting them as missing.
Make the mismatch failures stand out better by putting key
reason for the mismatch in all caps and ending with !!!