climate:  Aux heat appears to be deprecated.
The new climate UI introduced in HA 2023.9 no longer contains a switch
for Aux heat where it is supported.

Based on a rejected bug report in the frontend repository, this
appears to be deliberate, based on an architecture decision from July.
None of the other decisions from that architecture discussion have
been implemented, so there is no announcement about this deprecation,
and the backend documentation continues to advertise aux_heat has a
feature of the climate platform to both end users and developers.  But
since it is no longer accessible from the UI, it seems urgent to adapt
to the decision for the 4 devices that support the feature.

Also based on that architecture discussion, only one of the devices
actually implemented "aux heat", which is an automated feature on heat
pumps to start an auxilary heater when the temperature is too low for
the heat pump to function efficiently.  The other devices implement
"Emergency heat", which is a feature under user control to manually
switch on an electric element.  Names have been chosen as appropriate
based on that convention.

Discussion on PR #1331
Architecture discussion: https://github.com/home-assistant/architecture/discussions/932