Smardeer lock: rename request_unlock as there is no approve.
The request_unlock dp is designed to be paired with an approve_unlock
in this integration. But this device has only the request, no approve.
It is not clear how that is supposed to work, so just rename it to get it
out of the way.  This will mean there is no way to unlock this from
Home Assistant (but maybe unlock_app is intended for that?).

Issue #1181