Create morris_premier_plus_dehumidifier.yaml
Hello! This is just a clone of the klarstein_dryfy_pro_connect_dehumidifier.yaml file with minor changes for the Morris MDP-24400IW dehumidifier.
From Tuya API there are no more DPs returned, so they are basically identical. The only difference is that while the Klarstein one has the following modes: Auto/Boost/Eco/Purify, the Morris one has Auto/High/Low/Fan respectively.
So I just changed the relative values to the ones the Morris one uses. What I cannot find is where the Boost/Eco icons were coming from. So I simply add icons to the relevant section but I doubt they display.
This pull request is not actually createdwith the intention to be commited (unless you believe it is OK ) but rather in order to let you know that the Morris dehumidifier is working fine with the klarstein yaml file, except from those minor differences, so that you can perhaps add a "proper" version yourself.

P.S.: Thanks for your great work !
P.P.S.: The Boost icon (mdi: rocket-launch) is nowhere to be found in the component. How is it displayed?