Fan: add additional percentage and preset parameters to async_turn_on
Some time ago HA quietly added a fan specific turn_on method that takes
additional parameters.  In 2023.12 they have quietly broken the fallback
to the standard ToggleEntity turn_on that allowed this integration to keep
functioning.

This change is just to quickly resolve the errors that HA 2023.12
upgrade has caused for fan entities, it does not do anything with the
percentage and preset_mode parameters, that will need to be done in the
next release.

Issue #1380 (from discussion #1379)