Parkside charger: make dp 11 an int, dp 1 optional.
Based on user feedback, dp 11 seems to be a fault code (bit field), and
dp 1 is not always sent.  The original report did not include logs, and there
was never feedback from the original reporter.

dp 2 is a fixed string with the model name, so does not make sense to
have as a sensor. Moved to an attribute.

Issue #1387