Add support for VIVOSUN DE0003 Dehumidifier
This PR adds support for the VIVOSUN DE0003 Dehumidifier (also known as
the VIVOSUN 10L Dehumidifier on Amazon). Results from the Tuya API
showed that this dehumidifer was really similar to the JJPro JPD02
dehumidifier, so this config is mostly a copy of that (they even share
the same name for the continuous mode: Continuities).