Loratap curtain switch: split null out from normal conditions.
The code for handling multiple conditional values does not allow null to be
used in a list. There were reasons for that, so instead of changing the code
to work, change the config to split that condition out by itself.

May fix the case where buttons are not getting enabled, but users are reporting
that they are all OK if they simplify the config to only work for their
particular device.  Unit tests show that everything else about the
multi-conditional configs works as expected, so should not be any different
than the simple configs that are reportedly working.

Issue #1442, #1448, post-merge comments on PR #646