Motion detection: convert to event where dps were optional.
binary_sensor has a motion class, and some devices seem to always
return a value, so should remain as binary_sensors.

Only the following devices had the motion dp marked as optional, so
would be better represented as a stateless event:
- ble_solar_light
- pir_spotlight

In the case of the latter, the forced dp would create spurious events,
so removed the `force: true` and instead default the value to false
when no value has been reported yet.

Issue #1068