Setup: print useful errors, allow config to proceed on exceptions
When we get exceptions during entity creation, the exception was being
thrown to HA to handle. But this loses some of the info such as which
config file the error is in, and will also prevent other entities using
the same platform from being enabled.

Catch the exception to prevent other entities from being affected, and print
a useful error message. There is already a separate exception thrown if
no entities are available to set up for the platform, so that will still be
thrown if there are no other entities that would have been blocked by an
exception being thrown during initialisation.

Issue #1487