Smart watering device: remove unneeded mappings
Including mappings of true to ON and false to OFF is just confusing to anyone using this as a template for a new device, as yaml maps these to the same Python True and False values.  The values do not need to be mapped, and are not needed explicitly for a list of possible options, as a switch has fixed options.