Update config_flow unit tests for change in HA 2024.1.2
HA 2024.1.2 now includes minor_version in the config, even when the
integration did not set it.  minor_version was added in HA 2024.1.0,
so it is not set by this integration yet.

Update the unit tests for this change in HA behaviour, and update the
pytest-home-assistant-custom-component version to ensure HA 2024.1.2
or later is used in the tests.