TTEC vacuum: remap pause to activate and use command directly.
An activate dp is required to start/pause the vacuum if there is no
start and pause command in the command dp.  We had followed the tuya
naming of this as "pause" to integrate the pause command into the
command dp, but there is still no "start" command.  It is better to
separate it back out and reverse the logic to make it an "activate" dp.

Issue #1545