CI: change from black and isort to ruff.
Changes to black today broke CI, so this is a catalyst to make the
move to ruff, as Home Assistant already has.

A few issues picked up, and apparently isort has not been working
(or maybe was reporting its findings without flagging an error).

Also should do general linting, but there will be quite a bit of
cleanup for that.