TTEC vacuum: add separate buttons for first three dps
Leave the 1/2 combo for activate, but remove dp 3 as power, as this is
for sending the vacuum to charge.

Add the separate buttons in case the vacuum interface is not able to control
it (eg device does not like two commands sent at once).

Based on comment in #1393 about a similar vacuum with 3 booleans for power_go,
pause and switch_charge