Kogan bladeless fan: change time remaining to integer
Sensor with unit min requires an integer to work. Even if the actual data
is an integer in double quotes, integer as a type will still work.

PR #1480 comment:

I don't know why but combining the 2 fans, fan1csa and fansa "broke" the
devices. The became greyed out in Home Assistant. I re-added the seperated
yamls and they both worked again.

This seems to be the only possible blocker to them being detected correctly,
as the config is working as expected in tests.