Remove support for Tadiran products.
Although this company was split several years ago, distancing the
consumer air conditioning division from the military radio and other
military related divisions, there are still common board members
between the companies, so it appears to be a distinction purely to
avoid sanctions.

I cannot support a company that is so closely linked to genocide no
matter how widely available its devices are.